+++ src/components/Editor/EditorHeader/EditorHeader.scss
