+++ src/components/Editor/StatusBar/StatusBar.scss
