+++ src/components/Modal/AccountModal/AuthTokenPanel.js
