+++ src/components/Modal/AccountModal/Panel.scss
