+++ src/components/Modal/ErrorModal/ErrorModal.js
@@ -1,0 +1,5 @@
+const ErrorModal = () => {
+
+};
+
+export default ErrorModal;
